requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-google-genai
chromadb
couchbase
langchain-community